[TASK] Support meetingdata data on showAction
This requires a complete overhaul of the javascript file, as it needs to
interact a lot with smvplayer and jwplayer objects. So we're using this
change to also refactor the entire private javascript. The outcome will
be a neat javascript SVPS object that interacts with everything and
allows for relatively easy debugging.

The template needs adjustments to show the data via a meetingdata
object. We've decided to remove the script tag for javascript (as it
wouldn't be up to standards anyway), but we WILL utilize script tags for
json data that is going to be read by the SVPS.

For the above, we need to have a meetingdata object in the first place,
once that is a bit different than the REST response. Therefor, we need
to further expand the ResponseFactory to support options like json
conversion and mapping of properties.

The static TS needs to reflect these changes by adding the necessary
configuration.

Change-Id: I2bb2bbf6167dbca1e805d9a8149dc14623855838