[BUGFIX] Deactivate elements when smvplayer moves to next item
Speakers and topics are deactivated during onSeek() events that have no
matching timestamp, but they aren't when the player is allowed to move
to next() by itself.

The next() and previous() methods need to deactivate elements
automatically. This is a good time to introduce a SVPS.reset() method
which contains all code necessary once a jwplayer instance is destroyed.

While adding the reset method, improve some of the descriptions and
var-names of related properties and methods while we're at it.

Change-Id: Ie42be022a3db26a1bfda8f241ab416cecf0fb601