[BUGFIX] config.compressJs = 1 prevents players from finding files
With compressJs enabled, the player scripts are loaded from
typo3temp/compressor instead of from the provided dir. This stops the
player scripts from finding all of their dependencies out of the box.

By default, I'm going to disableCompression on the player scripts so
that this doesn't become a usability problem. Knowledgable users can
still re-enablle it and copy the necessary files to the relevant
directories.

Change-Id: Ibcb33253f8198daef6e659e32c28722963651fab