[TASK] Create eID script for polling meetingdata
The script should be straightforward: when called, it should simply
request the meetingdata as configured.

"As configured" being, as is imposed by the configuration active on the
page, and meetingdata for the sessionhash currently being viewed.

These latter two pieces of data need to retrieved, as they aren't
readily available in eID context. The plan is to supply them as GET
variables, thus the eID script can be built to depend on those as its
source.

Change-Id: I4e74dc53e13d84f39db50a323a33d2288578ae95