[BUGFIX] eID script fails due to inflexible reponsemapping
The responsemapping has some very specific, static features:
- removing '.result'
- always returning an array of objects, even when we expect only 1
response object

MeetingdataRepository has a workaround to this, by returning only the
first element of the array. But this makes the repository inflexible, as
we now notice when trying to retrieve a forced raw response for the eID
context.

We need to fully realize the flexibility of the response-mapping that is
implicitly promised through our intended use.

Change-Id: Idb8fb056b78e72268dcdc410ffaf33e2956fae01