[TASK] Simplify Library/Rest/MagicResponse
MagicResponse uses an array to store its properties, but it turns out
this is not json_encode() friendly, which is a pain.

Why not simply define public variables at runtime instead?
MagicResponse's only functional difference from StdClass would be its
__call() implementation. This way, the class becomes json_encode()
friendly, and it's a cleaner and more simple approach (considering the
class is only to be used during debugging anyway).

Change-Id: I64acc87af27ed608f0191b3596c74767d1274418