[BUGFIX] meetingdata eID script returns unusable timeline data
The data returned is the original JSON response. We explicitly require a
separation between timelines of speakers and of topics, hence this is
unusable.

For this to work, we need to disable the forceRawResponse setting on the
repository, and re-encode the response class into JSON. This means that
any JSON-encoded property (through responsemapping config) needs to
decoded again before the entire class is encoded.

Because the latter is suboptimal, why not enhance the eID utility method
for Extbase Bootstrap initialization: allow additional configuration to
be given. This way, we can overrule any REST TS setting that sets
properties.json = 1, thus preventing their initial re-encoding.

Change-Id: Ie02c3fa393adcfe25906a005fd8ce638b281e018