[TASK] SvpStarter should process livestream meetingdata
The new method should be publicly available, thus be a part of the SVPS
object. This way, it can be called from the SVPP object.

The method receives as input a JSON object, already parsed, as that is
what jQuery.get() returns.

Because the method is publicly available, the method can be tested
thoroughly via console, before really activating the polling feature.

Also, the requirements for this method clearly show that
initEventHandlers() is not useful for livestreams. The entire timeline
data as it is delivered through the template probably isn't useful
either, as we only really need to know the last timestamp for both
speaker and topic.

Change-Id: If7924bd1ce1de4bb6e2e2de143242fb1abf8f992