[TASK] Provide meaningful error early on, on no connection
When the REST response is invalid or failed, we only find out when an
exception is thrown because of an unexpected response structure, and
that message isn't meaningful yet.

There should be an exception thrown right when we've sent the request,
so that processing stops as soon as can be established there is no
connection. Also, utilize internal error-reporting mechanisms to
retrieve the relevant debugging error message. This can be replaced
later on with a frontend-friendly one, when error-reporting in general
is vastly improved.

Change-Id: Id5d5bc696a2080faf7c945e4c36cbb4823765d1e