[BUGFIX] Sometimes no timestamps returned on polling after page-refresh
The PHP session survives a page refresh or simply some navigating away
and returning later on within the same browser session. However, if
during that time, the meetingdata from the eID script is unchanged, no
timestamps will be returned to indicate current active elements.

One way to solve this, is by using the initial poll() call in the SVPP
object to indicate a forced return value to the eID script. The eID
script could listen to a 'force' GET var and determine by its value if
the session hash should be compared with the current meetingdata.
'Force' would then indicate that it is forced to return meetingdata,
regardless of session values.

The initial poll() script is also called when polling is re-initialized,
e.g. during a pause and then play, or simply when refreshing or
revisiting the page within the same browser session.

Change-Id: I0505365737f013132429c468e0275a4a16bb70fc