[TASK] Add onPlay & onPause to reattach-events, fix qual/lang methods
Livestream testing exposed a problem with choosing a different quality
or language: these also destroy the jwplayer object, just like next()
and previous() do. So those methods need to be overruled as well.

During a livestream, the onPause and onPlay events used for polling need
to be reattached when a jwplayer-reset happens. This can be done by
applying the same techniques as I've done to onTime and onSeek.

Change-Id: Icce70b7d70d4f0b0b75a79230f9851565d7bb533