[TASK] Create a sort VH for arrays
It should be possible to define its order and a propertypath to sort by.

In context: We want to be able to sort an array of events by
event.startUnix, descending.

Ideally, the sorted array is usable for anything afterwards. This means
the sort VH can't be an extended for.each VH.

Change-Id: I4c131e86b7623ed9765ae14bd267a937cdd9da11