[TASK] Create configurable AssetProvider Lib
The AssetProvider needs to be easily extendable and highly configurable.
It's primary purpose is to be able to provide assets only on pages
containing this extension's plugins. This is especially useful because
of the loads of javascript required by streamovations_vp only during
Video->showAction.

- jQuery
- jwPlayer
- smvPlayer
- SVPS
- SVPP

Although jQuery might be included on a site on any page, the others are
usually hardly necessary on most pages. Also, to get initialization to
run only on the plugin pages, I would normally have to:
- place SVPS.init() directly in the template, which goes against WCAG
- or, rely on Vhs' Asset VHs, but I dont want it as dependency

Also, I don't like that Vhs' Asset VHs are placed in the template. Of
course it's great that get an added flexibility, considering you can
have different templates on different pages. But you'd need to configure
this via TS anyway. Why not provide asset configuration via TS directly,
in a way that is mostly true to actual TS reference?

Lo and behold, the AssetProvider Lib was born. And is properly cached
even, through no special steps which only adds to any
instability TYPO3 CMS might have, as vhs, flux, fed, etc. have
demonstrated.

Oh, and get rid of obsolete jQuery lib. (accidentally committed)

Change-Id: I620193e1794753a3f8fa23945b442e5c85c59fc7