[TASK] Implement interfaces on abstract classes
I've been implementing interfaces on classes, which is fine. But the
same classes extend an abstract which is meant to complement any class
implementing the interface.

In those cases, it's cleaner to define the implementation on the
abstract.

RestRepository\ResponseFactory wasn't a singleton yet, but becomes one
thorugh this change.

Both concrete factory classes in RestRepository lib implement their own
interface, because this is different for both, regardless of both
extending FactoryAbstract.

Change-Id: I3dfea28a6801bbdad224d599fb2465b882b6b914