[BUGFIX] speaker image dir EXT prefix is not always resolved
When a a speaker image dir or a the default image is prefixed with EXT:,
this isn't resolved in the speaker template nor in the SVPS javascript
object.

For the javascript object, we need to use a TS UserFunc.

For the template, we could simply change the img tag to an f:image tag.

Change-Id: I7a6f8ad0f3c2bac98e4d26d10bcbabd539ced10d