[TASK] Expand MagicResponse for completeness
In preparation of creating a responseMapper solution, we need the
MagicResponse to be a full alternative. At this point, every
multi-dimensional array is returned as just that: an array. But when we
move to a full reponse mapper with actual domain models, arrays need to
be objectStorages and such, for consistency with other types of domain
models. The effect is that treating them as an array as you would with
MagicResponse, will completely fail with the response mapper enabled,
thus leaving the 2 methods un-interchangable.

A way to solve this, is to make MagicResponse mimic a domain model more
accurately. When returning a value that is an array, we could do an
attempt of determining whether this should have been an object (simply
returned as a new MagicResponse), or an objectStorage containing
MagicResponses.

When this logic is added to MagicResponse, we're left with one more
issue: what if a property of a response is to be treated as an array,
according to variable annotations as defined in the domain model? If the
array is returned as an objectStorage, this isn't that big of deal,
because an objectStorage implements several PHP interfaces that allow it
to be treated as an array. Countable, Iterator and ArrayAccess come to
mind. Suppose MagicResponse _also_ implements these interfaces, then any
value orignally an array but returned as an object, could still be
approached as an array in the most important cases!

Change-Id: I60c0fd6ef42ea060541aca9331f95d78aff18b8f