[FEATURE] Create REST ResponseMapper
The ResponseMapper should be run by ResponseFactory if extension TS
config rest.features.disableResponseMapper != 1.

We can take inspiration from the Extbase DataMapper class, as it does
more or less the same with objects reconstituted from the database while
only given their ID.

The mapper needs to know the object type as a class. Currently, the
repository only stores the shorthand, but not the entire class and
namespace. This needs to change, and both factories need to determine
the shorthand from the new classname version of objectType. Also the
shorthand can have an uppercase first character, for consistency with my
other libraries.

Change-Id: I2d90062187d292aacf5ad3117df075c708b96555