[FEATURE] Cache REST Requests
I have decided not to cache the plugins. Instead, I will cache the rest
requests, and this will be configurable per rest-repository via TS.

This way, a user could easily alter configuration to suit his needs per
request-type. This also enables us to configure the meetingdata requests
for the eID script specifically.

For now the only supported settings are:
- lifetime (in seconds, positive integer)
- disable (0 (=default) or 1)

Change-Id: I6304ba5daf2f1011ffc3c63bbbe60530f294d462