[TASK] Dont throw exception if qualities is missing in response
Currently, an exception is thrown if 'qualities' is missing in a
playlist response, but only due to a check in the PlaylistService.

Not only can we currently assume the REST response is valid, this check
really is in the wrong place. It's better to leave this to a Validator,
once RestRepository supports this.

Change-Id: I299e599f33b6c25bb724ba684248c8bee78eed29