[TASK] Add TypoScript-driven AdvancedShowAction
The purpose of this action is to have a more flexible solution with
which we can configure which video to view, based on e.g. external
parameters. With TypoScript, we can provide i.e. the category and date
parameters for the video by taking another extension's GET parameters
and/or resolving ID's to database-values with the CONTENT TypoScript
content object.

This allows for the streamovations plugin to be influenced by any other
plugin, like when you want to show a streamovations-video with a
specific plugin's content that is of the same subject as the
streamovations-video, with the right TypoScript configuration.

The action is called "advanced", because you need to know your way with
TypoScript to make use of it. Of course, you can still provide static
values to the streamovations parameters as well, not utilizing any
TypoScript content object. This also provides the use-case where you
simply want to configure the plugin via TypoScript-set values.

The typoscript settings that are currently supported are:

tx_streamovationsvp.settings.advanced {
  category
  subCategory
  tags
  dateAt = note the distinction with the next
  dateTimeAt = noet the distinction with the previous
  enable = this boolean switch needs to be true (= 1)
}

Change-Id: I75d740410d71f6dc5523dfcbbce8e308dc9f0282