[BUGFIX] SVPS:play trigger does not register since new SMV player
The new SMV player changes the player div id, which has the unnoticed
side-effect of the SVPS:play trigger not being registered to the actual
player div.

An easy way to fix this, is by using an additional container div whose
id/classes don't get changed by SMV player, and register the triggers on
that instead.

Change-Id: I276ac0d6f5381d8d8acd61b3525ddd1947d98ac7