[BUGFIX] Livestream plugin breaks if VOD is simultaneously active
The issue occurs if a VOD has some magical (unrealistic) end time set
after the current livestream start time. In these situations, the
LiveStream plugin finds the VOD instead of the Livestream, assuming
there can ever be only 1 stream active @ now time.

This could be fixed by a simple search for the first occurring
livestream.

Note that this issue shouldn't really occur in production, but it
happened during testing, so it needs to be fixed either way.

Change-Id: I2f0205cca7fe1c618cb9d39ed30257d243169a2e