[TASK] Support REST Cache tags for debugging purposes
Cache tags should be disabled by default, as they add overhead and their
only use to me is for debugging. When they're enabled via the rest
repository cache setting, they should register all elements that make up
a different cache entry:

- repository name
- raw output yes/no
- type of output (e.g. json/xml)
- cache lifetime?

A REST request only has the object type, which we already have converted
to a repository name in the factories.. so what if we map and store that
conversion in a RepositoryMapper singleton which we can use both in the
request AND factories?

Change-Id: I003e9ad9437c709b7ddac8426c66da2ce78d4bfe