[TASK] Allow caching of polling meetingdata
With caching disabled on polling, we're ensured of always having an up
to date meetingdata response. But in production, this can lead to
literally thousands or tenthousands or.. even more concurrent requests
to the Streamovations REST API, per polling interval!

Enabling caching will lighten this load. However, since the showAction
already gets and stores the Meetingdata response with a lifetime of 3600
seconds, this is the same response the pollingdata will get. To remedy
this, we should let a cache entry be identified additionally by the
intended lifetime of said entry. Doing so will result in different cache
entries for meetingdata: one for showAction, and one for polling.

Aditionally, we have to keep in mind the polling interval. If a visitor
is presented with the the cached response he's seen earlier, then this
would defeat the purpose of the chosen polling interval.

Change-Id: I21c121ad09f2600965c226fdfe3a74a12e4e42b8