[BUGFIX] Livestream caches too long
Since livestream utilizes the same repositories as show, as well as an
additional Event check, it is subject to the same caching rules. That
means the default lifetime of 3600 seconds applies. This would mean that
if the Event check is cached 5 minutes before a livestream starts, it
wouldn't show up until it's already 55 minutes in. We've seen this
happen during tests.

To fix this, we need to be able to set different repository settings per
controller/action, like how assetProvider sets these. While doing so, we
need to be mindful of liveStreamAction's forward() call to showAction,
and the trouble of how it could result in resetting controller/action
parameters for the repository settings.

All of this requires a refactoring of the RestRepository library's
settings management. Doing so will result in a change for setting
features: they will henceforth need to be set per repository (or just in
default).

Change-Id: I184b39f8b98a4a4652f3ecaf53807ff2e5780a21