[TASK] Create single API for SVPS getState() method
Doing an if statement on every seek() is just bad practice. Instead, we
should define a getState() method on the smvplayer object to unify its
API with the jwplayer object.

While doing so, note that there is no logging if there ever is an issue
from adding new methods to the smvplayer object. Any future issue risen
from this would be hard to debug. So also do some checks when adding
methods, and log if it already exists.

Change-Id: I7bfbd482023dd8f989aeb18c8ad9771813da5953