[TASK] Add support for TYPO3 8.7.11+ Guzzle-changes
TYPO3's getUrl() method was changed in 8.7.11 in how it populates the
report variable. No longer is status stored in 'error' for GET 200s.
Instead, it remains 0.

Instead of allowing a dependency to such arbitrary changes without
mention, I'm replacing our use of getUrl() with a more direct usage of
Guzzle.

Change-Id: I089efec741597f57d3c32914776d1a2eba041068